+++ R/standardizedSolution_boot_ci.R
@@ -76,7 +76,7 @@
 #' with diagnostic functions such
 #' as [plot_boot()].
 #'
-#' @param object A [lavaan-class]
+#' @param object A 'lavaan'-class
 #' object, fitted with 'se = "boot"'.
 #'
 #' @param level The level of confidence
@@ -108,6 +108,13 @@
 #' estimate to (b) the distance of the
 #' delta-method limit from the point
 #' estimate. Default is `FALSE`.
+#'
+#' @param boot_ci_type The type of the
+#' bootstrapping confidence intervals.
+#' Support percentile confidence intervals
+#' (`"perc"`, the default) and
+#' bias-corrected confidence intervals
+#' (`"bc"` or `"bca.simple"`).
 #'
 #' @param ... Other arguments to be
 #' passed to
@@ -159,6 +166,10 @@
 #' # and standardized solution
 #' print(std, output = "text", standardized_only = FALSE)
 #'
+#' # plot_boot() can be used to examine the bootstrap estimates
+#' # of a parameter
+#' plot_boot(std, param = "ab")
+#'
 #' @name standardizedSolution_boot_ci
 NULL
 
@@ -171,6 +182,7 @@ standardizedSolution_boot_ci <- function(object,
                                          save_boot_est_std = TRUE,
                                          force_run = FALSE,
                                          boot_delta_ratio = FALSE,
+                                         boot_ci_type = c("perc", "bc", "bca.simple"),
                                          ...) {
     if (!inherits(object, "lavaan")) {
         stop("The object must be a lavaan-class object.")
@@ -189,15 +201,20 @@ standardizedSolution_boot_ci <- function(object,
                       t = out_all,
                       R = nrow(out_all))
     # Adapted from boot
-    boot_ci <- sapply(seq_along(est_org), function(x) {
+    boot_ci <- sapply(seq_along(est_org), function(x,
+                                                   boot_type = boot_ci_type) {
                           if (isTRUE(all.equal(stats::var(out_all[, x], na.rm = TRUE), 0)) ||
                               all(is.na(out_all[, x]))) {
                               return(c(NA, NA))
                             }
-                          boot::boot.ci(boot_tmp,
-                                index = x,
-                                type = "perc",
-                                conf = level)$percent[4:5]
+                          # boot::boot.ci(boot_tmp,
+                          #       index = x,
+                          #       type = "perc",
+                          #       conf = level)$percent[4:5]
+                          boot_ci_internal(t0 = est_org[x],
+                                           t = out_all[, x],
+                                           level = level,
+                                           boot_type = boot_type)
                         })
     boot_ci <- t(boot_ci)
     colnames(boot_ci) <- c("boot.ci.lower", "boot.ci.upper")
@@ -216,6 +233,7 @@ standardizedSolution_boot_ci <- function(object,
       }
     class(out_final) <- c("std_solution_boot", class(out))
     if (save_boot_est_std) {
+        colnames(out_all) <- std_names(object, ...)
         attr(out_final, "boot_est_std") <- out_all
       }
     fit_summary <- lavaan::summary(object)
@@ -379,4 +397,65 @@ std_names <- function(object, ...) {
                         group.partial = "",
                         type = "user")
     return(std1$lavlabel)
+  }
+
+#' Internal function to form bootstrap CI
+#' @noRd
+
+# Internal function for different types
+# of bootstrap CI.
+# Only work for one statistic.
+# Adapted from manymome
+
+boot_ci_internal <- function(t0,
+                             t,
+                             level = .95,
+                             boot_type = c("perc", "bc", "bca.simple")) {
+    boot_type <- match.arg(boot_type)
+    out <- switch(boot_type,
+                  perc = boot_ci_perc(t0 = t0, t = t, level = level),
+                  bc = boot_ci_bc(t0 = t0, t = t, level = level),
+                  bca.simple = boot_ci_bc(t0 = t0, t = t, level = level))
+    out
+  }
+
+# Adapted from manymome
+#' @noRd
+
+boot_ci_perc <- function(t0,
+                         t,
+                         level = .95) {
+    boot_tmp <- list(t0 = t0,
+                     t = matrix(t, ncol = 1),
+                     R = length(t))
+    ci <- boot::boot.ci(boot_tmp,
+                        type = "perc",
+                        conf = level)
+    ci$perc[4:5]
+  }
+
+# Adapted from manymome
+#' @noRd
+
+boot_ci_bc <- function(t0,
+                       t,
+                       level = .95) {
+    # Compute the bias correction
+    z0_hat <- stats::qnorm(mean(t < t0,
+                                na.rm = TRUE))
+    z_alpha_l <- stats::qnorm((1 - level) / 2)
+    z_alpha_u <- -z_alpha_l
+    # BC percentiles
+    bc_a_l <- stats::pnorm(2 * z0_hat + z_alpha_l)
+    bc_a_u <- stats::pnorm(2 * z0_hat + z_alpha_u)
+    # Corresponding confidence level for percentile CIs
+    bc_conf_l <- 1 - bc_a_l * 2
+    bc_conf_u <- 1 - (1 - bc_a_u) * 2
+    boot_ci_bc_l <- boot_ci_perc(t0 = t0,
+                                 t = t,
+                                 level = bc_conf_l)
+    boot_ci_bc_u <- boot_ci_perc(t0 = t0,
+                                 t = t,
+                                 level = bc_conf_u)
+    c(boot_ci_bc_l[1], boot_ci_bc_u[2])
   }